Stage 3a: move the money rules to the server (DAT-2, DAT-6, ARC-5 partial)
The rounding rule lived only in JavaScript and the server stored whatever
amount the browser sent, so two screens could disagree about the same sale and
both were accepted. MoneyRules now owns the rule on the server, in BigDecimal,
and the client's figure is treated as a checksum: a bulk entry whose submitted
amount disagrees with the server's calculation is rejected rather than silently
overwritten.

The bird invariant (loaded = sold + mortality + returned) is now enforced
server-side too. It was previously only checked in the React form, so driver-app
and direct API entries could write loads that never reconcile.

Replaces the untyped List<Map<String, Object>> the bulk endpoint accepted with
SaleLineDto. The map required each field to arrive as a particular JSON type -
kilograms, rate, payment and birds as strings, amount and pending as numbers -
and threw on the other one. Two live 500s came from this:

  - a row with weight and rate but untouched birds sent birds as the number 0,
    and (String) 0 threw ClassCastException
  - typing a payment and then clearing it sent "", and Integer.parseInt("")
    threw NumberFormatException

Jackson now coerces either JSON form, and accept-empty-string-as-null-object
turns a cleared input into null, which the money rules read as zero.

Balances are updated from the rows actually saved rather than from the request,
since pending is now recalculated server-side.

Frontend: the on-screen totals summed every grid row while submission sent only
complete ones, so a row with birds but no weight was counted on screen and not
sent. With the server reconciling against the lines it receives, that mismatch
would have rejected valid entries. Both now use one shared isCompleteSaleLine
predicate.

Tests: 20 on the server, including a parity test that replays 126 amounts
so the two cannot drift apart without a build failure. 42 on the frontend.

Also corrects CustomerRepository.updateBalanceAmount, which took an Integer
customerId while Customer.id is Long.

Money and weight column types are unchanged; that migration is stage 3b.

+++ src/main/java/com/app/repository/CustomerRepository.java
@@ -19,7 +19,7 @@ public interface CustomerRepository extends JpaRepository<Customer, Long> {
 	
 	@Modifying
     @Query("UPDATE Customer c SET c.balanceAmount = c.balanceAmount + :amount WHERE c.id = :customerId")
-    int updateBalanceAmount(@Param("customerId") Integer customerId, @Param("amount") Integer amount);
+    int updateBalanceAmount(@Param("customerId") Long customerId, @Param("amount") Integer amount);
 
 }
